+++ src/test/java/de/samply/qualityreportgenerator/QualityReportGeneratorApplicationTests.java
@@ -1,8 +1,10 @@
 package de.samply.qualityreportgenerator;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+@Disabled
 @SpringBootTest
 class QualityReportGeneratorApplicationTests {